backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
sqlalchemy>=2.0.25
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pdfplumber>=0.11.0
pydantic>=2.5.0
python-dotenv>=1.0.0
dashscope>=1.23.4
PyYAML>=6.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
chromadb>=0.5.0
pysqlite3-binary>=0.5.0
